Fix test filename typo, add .idea (phpstorm project folder) to gitignore, Add new method in OnboardingManager "nextUnfinishedStep" which returns the next unfinished step or null if all is completed, added a middleware example in the README.
